requirements.txt
 streamlit
openai-whisper
transformers
torch
nltk
scikit-learn
fpdf